Исправление ошибок CheckStyle в Task.java
